Revert "featureMoviImg"
